Share ArmorFlags
