+++ SharedSE/SharedSE.vcxitems.filters
@@ -382,6 +382,7 @@
     <ClInclude Include="$(MSBuildThisFileDirectory)TES3BodyPartFlags.h">
       <Filter>Header Files\TES3</Filter>
     </ClInclude>
+    <ClInclude Include="$(MSBuildThisFileDirectory)TES3ArmorFlags.h" />
   </ItemGroup>
   <ItemGroup>
     <ClCompile Include="$(MSBuildThisFileDirectory)NIMatrix33.cpp">
